Remove ResultBearing, and rename ResultSetIterable back to ResultBearing
